requirements.txt
fastapi
gunicorn
uvicorn
jinja2
pytest
requests
httpx
pre-commit
python-decouple>=3.8
aiofiles
python-multipart>=0.0.6
pillow
pytesseract
PyJWT>=2.8.0
passlib>=1.7.4
bcrypt>=4.0.1